+++ CHANGELOG.md
@@ -13,12 +13,13 @@ and this project adheres to
 - Add `wait` methods back
 - Partial single recv `WinDivert::partial_recv()`
 - `WinDivertSendError`
+- Add `WinDivert<()>::install(path: &Path)`
 
 ### Changed
 
 - `WinDivertError` has a `Send` variant
 - `WinDivert::recv()` and `WinDivert::recv_ex()` buffer made mandatory on data
-  capturing layers and removed from non capturing layers.
+  capturing layers and removed from non-capturing layers.
 - Bumped `windows-rs` to 0.51.1
 - Code refactor and cleanup